requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-Marshmallow
Flask-Cors
gunicorn